Extract ShowQueriesRewrite from StatementAnalyzer
It is not intuitive that StatementAnalyzer does rewrites and analysis.
This rewrite extract rewrite responsibility to separate entity.
Thanks to that StatementAnalyzer becomes easier to be understood.
