+++ presto-main/src/main/java/com/facebook/presto/sql/analyzer/Analyzer.java
@@ -17,8 +17,10 @@ import com.facebook.presto.Session;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.security.AccessControl;
 import com.facebook.presto.sql.parser.SqlParser;
+import com.facebook.presto.sql.rewrite.StatementRewrite;
 import com.facebook.presto.sql.tree.Expression;
 import com.facebook.presto.sql.tree.FunctionCall;
+import com.facebook.presto.sql.tree.Node;
 import com.facebook.presto.sql.tree.Statement;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
@@ -55,9 +57,10 @@ public class Analyzer
 
     public Analysis analyze(Statement statement)
     {
+        Statement rewrittenStatement = StatementRewrite.rewrite(session, metadata, sqlParser, statement);
         Analysis analysis = new Analysis();
         StatementAnalyzer analyzer = new StatementAnalyzer(analysis, metadata, sqlParser, accessControl, session, experimentalSyntaxEnabled, queryExplainer);
-        RelationType outputDescriptor = analyzer.process(statement, new AnalysisContext());
+        RelationType outputDescriptor = analyzer.process(rewrittenStatement, new AnalysisContext());
         analysis.setOutputDescriptor(outputDescriptor);
         return analysis;
     }
