+++ presto-main/src/main/java/com/facebook/presto/sql/analyzer/StatementAnalyzer.java
@@ -17,11 +17,7 @@ import com.facebook.presto.Session;
 import com.facebook.presto.metadata.FunctionKind;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.metadata.QualifiedObjectName;
-import com.facebook.presto.metadata.SessionPropertyManager.SessionPropertyValue;
-import com.facebook.presto.metadata.SqlFunction;
 import com.facebook.presto.metadata.TableHandle;
-import com.facebook.presto.metadata.TableLayout;
-import com.facebook.presto.metadata.TableLayoutResult;
 import com.facebook.presto.metadata.TableMetadata;
 import com.facebook.presto.metadata.ViewDefinition;
 import com.facebook.presto.security.AccessControl;
@@ -29,12 +25,8 @@ import com.facebook.presto.security.AllowAllAccessControl;
 import com.facebook.presto.security.ViewAccessControl;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ColumnMetadata;
-import com.facebook.presto.spi.ConnectorTableMetadata;
-import com.facebook.presto.spi.Constraint;
 import com.facebook.presto.spi.PrestoException;
-import com.facebook.presto.spi.SchemaTableName;
 import com.facebook.presto.spi.security.Identity;
-import com.facebook.presto.spi.session.PropertyMetadata;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeSignature;
 import com.facebook.presto.sql.ExpressionUtils;
@@ -47,17 +39,12 @@ import com.facebook.presto.sql.planner.Symbol;
 import com.facebook.presto.sql.planner.optimizations.CanonicalizeExpressions;
 import com.facebook.presto.sql.tree.AliasedRelation;
 import com.facebook.presto.sql.tree.AllColumns;
-import com.facebook.presto.sql.tree.ArrayConstructor;
-import com.facebook.presto.sql.tree.BooleanLiteral;
-import com.facebook.presto.sql.tree.Cast;
 import com.facebook.presto.sql.tree.ComparisonExpression;
-import com.facebook.presto.sql.tree.CreateTable;
 import com.facebook.presto.sql.tree.CreateTableAsSelect;
 import com.facebook.presto.sql.tree.CreateView;
 import com.facebook.presto.sql.tree.DefaultTraversalVisitor;
 import com.facebook.presto.sql.tree.Delete;
 import com.facebook.presto.sql.tree.DereferenceExpression;
-import com.facebook.presto.sql.tree.DoubleLiteral;
 import com.facebook.presto.sql.tree.Except;
 import com.facebook.presto.sql.tree.Explain;
 import com.facebook.presto.sql.tree.ExplainFormat;
@@ -67,7 +54,6 @@ import com.facebook.presto.sql.tree.Expression;
 import com.facebook.presto.sql.tree.FieldReference;
 import com.facebook.presto.sql.tree.FrameBound;
 import com.facebook.presto.sql.tree.FunctionCall;
-import com.facebook.presto.sql.tree.GroupBy;
 import com.facebook.presto.sql.tree.GroupingElement;
 import com.facebook.presto.sql.tree.Insert;
 import com.facebook.presto.sql.tree.Intersect;
@@ -75,7 +61,6 @@ import com.facebook.presto.sql.tree.Join;
 import com.facebook.presto.sql.tree.JoinCriteria;
 import com.facebook.presto.sql.tree.JoinOn;
 import com.facebook.presto.sql.tree.JoinUsing;
-import com.facebook.presto.sql.tree.LikePredicate;
 import com.facebook.presto.sql.tree.LongLiteral;
 import com.facebook.presto.sql.tree.NaturalJoin;
 import com.facebook.presto.sql.tree.Node;
@@ -88,21 +73,10 @@ import com.facebook.presto.sql.tree.Row;
 import com.facebook.presto.sql.tree.SampledRelation;
 import com.facebook.presto.sql.tree.SelectItem;
 import com.facebook.presto.sql.tree.SetOperation;
-import com.facebook.presto.sql.tree.ShowCatalogs;
-import com.facebook.presto.sql.tree.ShowColumns;
-import com.facebook.presto.sql.tree.ShowCreate;
-import com.facebook.presto.sql.tree.ShowFunctions;
-import com.facebook.presto.sql.tree.ShowPartitions;
-import com.facebook.presto.sql.tree.ShowSchemas;
-import com.facebook.presto.sql.tree.ShowSession;
-import com.facebook.presto.sql.tree.ShowTables;
-import com.facebook.presto.sql.tree.SimpleGroupBy;
 import com.facebook.presto.sql.tree.SingleColumn;
 import com.facebook.presto.sql.tree.SortItem;
 import com.facebook.presto.sql.tree.Statement;
-import com.facebook.presto.sql.tree.StringLiteral;
 import com.facebook.presto.sql.tree.Table;
-import com.facebook.presto.sql.tree.TableElement;
 import com.facebook.presto.sql.tree.TableSubquery;
 import com.facebook.presto.sql.tree.Unnest;
 import com.facebook.presto.sql.tree.Use;
@@ -126,7 +100,6 @@ import com.google.common.primitives.Ints;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.IdentityHashMap;
 import java.util.Iterator;
@@ -135,50 +108,23 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 
-import static com.facebook.presto.connector.informationSchema.InformationSchemaMetadata.TABLE_COLUMNS;
-import static com.facebook.presto.connector.informationSchema.InformationSchemaMetadata.TABLE_INTERNAL_PARTITIONS;
-import static com.facebook.presto.connector.informationSchema.InformationSchemaMetadata.TABLE_SCHEMATA;
-import static com.facebook.presto.connector.informationSchema.InformationSchemaMetadata.TABLE_TABLES;
-import static com.facebook.presto.execution.SqlQueryManager.unwrapExecuteStatement;
 import static com.facebook.presto.metadata.FunctionKind.AGGREGATE;
 import static com.facebook.presto.metadata.FunctionKind.APPROXIMATE_AGGREGATE;
 import static com.facebook.presto.metadata.FunctionKind.WINDOW;
-import static com.facebook.presto.metadata.MetadataUtil.createQualifiedName;
 import static com.facebook.presto.metadata.MetadataUtil.createQualifiedObjectName;
 import static com.facebook.presto.spi.StandardErrorCode.INVALID_FUNCTION_ARGUMENT;
-import static com.facebook.presto.spi.StandardErrorCode.INVALID_TABLE_PROPERTY;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
-import static com.facebook.presto.sql.QueryUtil.aliased;
-import static com.facebook.presto.sql.QueryUtil.aliasedName;
-import static com.facebook.presto.sql.QueryUtil.aliasedNullToEmpty;
-import static com.facebook.presto.sql.QueryUtil.ascending;
-import static com.facebook.presto.sql.QueryUtil.caseWhen;
-import static com.facebook.presto.sql.QueryUtil.equal;
-import static com.facebook.presto.sql.QueryUtil.functionCall;
-import static com.facebook.presto.sql.QueryUtil.logicalAnd;
-import static com.facebook.presto.sql.QueryUtil.nameReference;
-import static com.facebook.presto.sql.QueryUtil.ordering;
-import static com.facebook.presto.sql.QueryUtil.row;
-import static com.facebook.presto.sql.QueryUtil.selectAll;
-import static com.facebook.presto.sql.QueryUtil.selectList;
-import static com.facebook.presto.sql.QueryUtil.simpleQuery;
 import static com.facebook.presto.sql.QueryUtil.singleValueQuery;
-import static com.facebook.presto.sql.QueryUtil.subquery;
-import static com.facebook.presto.sql.QueryUtil.table;
-import static com.facebook.presto.sql.QueryUtil.unaliasedName;
-import static com.facebook.presto.sql.SqlFormatter.formatSql;
 import static com.facebook.presto.sql.analyzer.ExpressionAnalyzer.createConstantAnalyzer;
 import static com.facebook.presto.sql.analyzer.ExpressionAnalyzer.getExpressionTypes;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.AMBIGUOUS_ATTRIBUTE;
-import static com.facebook.presto.sql.analyzer.SemanticErrorCode.CATALOG_NOT_SPECIFIED;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.COLUMN_NAME_NOT_SPECIFIED;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.COLUMN_TYPE_UNKNOWN;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.DUPLICATE_COLUMN_NAME;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.DUPLICATE_RELATION;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.INVALID_ORDINAL;
-import static com.facebook.presto.sql.analyzer.SemanticErrorCode.INVALID_SCHEMA_NAME;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.INVALID_WINDOW_FRAME;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.MISMATCHED_COLUMN_ALIASES;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.MISMATCHED_SET_COLUMN_TYPES;
@@ -191,7 +137,6 @@ import static com.facebook.presto.sql.analyzer.SemanticErrorCode.NESTED_WINDOW;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.NON_NUMERIC_SAMPLE_PERCENTAGE;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.NOT_SUPPORTED;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.ORDER_BY_MUST_BE_IN_SELECT;
-import static com.facebook.presto.sql.analyzer.SemanticErrorCode.SCHEMA_NOT_SPECIFIED;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.TABLE_ALREADY_EXISTS;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.TYPE_MISMATCH;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.VIEW_ANALYSIS_ERROR;
@@ -199,8 +144,6 @@ import static com.facebook.presto.sql.analyzer.SemanticErrorCode.VIEW_IS_STALE;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.VIEW_PARSE_ERROR;
 import static com.facebook.presto.sql.analyzer.SemanticErrorCode.WILDCARD_WITHOUT_FROM;
 import static com.facebook.presto.sql.planner.ExpressionInterpreter.expressionOptimizer;
-import static com.facebook.presto.sql.tree.BooleanLiteral.FALSE_LITERAL;
-import static com.facebook.presto.sql.tree.BooleanLiteral.TRUE_LITERAL;
 import static com.facebook.presto.sql.tree.ComparisonExpression.Type.EQUAL;
 import static com.facebook.presto.sql.tree.ExplainFormat.Type.TEXT;
 import static com.facebook.presto.sql.tree.ExplainType.Type.DISTRIBUTED;
@@ -210,21 +153,15 @@ import static com.facebook.presto.sql.tree.FrameBound.Type.FOLLOWING;
 import static com.facebook.presto.sql.tree.FrameBound.Type.PRECEDING;
 import static com.facebook.presto.sql.tree.FrameBound.Type.UNBOUNDED_FOLLOWING;
 import static com.facebook.presto.sql.tree.FrameBound.Type.UNBOUNDED_PRECEDING;
-import static com.facebook.presto.sql.tree.ShowCreate.Type.TABLE;
-import static com.facebook.presto.sql.tree.ShowCreate.Type.VIEW;
 import static com.facebook.presto.sql.tree.WindowFrame.Type.RANGE;
 import static com.facebook.presto.type.UnknownType.UNKNOWN;
 import static com.facebook.presto.util.ImmutableCollectors.toImmutableList;
 import static com.facebook.presto.util.Types.checkType;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.base.Strings.nullToEmpty;
 import static com.google.common.collect.Iterables.getLast;
-import static com.google.common.collect.Iterables.getOnlyElement;
 import static com.google.common.collect.Iterables.transform;
-import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
-import static java.util.stream.Collectors.toList;
 
 class StatementAnalyzer
         extends DefaultTraversalVisitor<RelationType, AnalysisContext>
@@ -255,399 +192,9 @@ class StatementAnalyzer
     }
 
     @Override
-    protected RelationType visitShowTables(ShowTables showTables, AnalysisContext context)
-    {
-        String catalogName = session.getCatalog().orElse(null);
-        String schemaName = session.getSchema().orElse(null);
-
-        Optional<QualifiedName> schema = showTables.getSchema();
-        if (schema.isPresent()) {
-            List<String> parts = schema.get().getParts();
-            if (parts.size() > 2) {
-                throw new SemanticException(INVALID_SCHEMA_NAME, showTables, "Too many parts in schema name: %s", schema.get());
-            }
-            if (parts.size() == 2) {
-                catalogName = parts.get(0);
-            }
-            schemaName = schema.get().getSuffix();
-        }
-
-        if (catalogName == null) {
-            throw new SemanticException(CATALOG_NOT_SPECIFIED, showTables, "Catalog must be specified when session catalog is not set");
-        }
-        if (schemaName == null) {
-            throw new SemanticException(SCHEMA_NOT_SPECIFIED, showTables, "Schema must be specified when session schema is not set");
-        }
-
-        if (!metadata.listSchemaNames(session, catalogName).contains(schemaName)) {
-            throw new SemanticException(MISSING_SCHEMA, showTables, "Schema '%s' does not exist", schemaName);
-        }
-
-        Expression predicate = equal(nameReference("table_schema"), new StringLiteral(schemaName));
-
-        Optional<String> likePattern = showTables.getLikePattern();
-        if (likePattern.isPresent()) {
-            Expression likePredicate = new LikePredicate(nameReference("table_name"), new StringLiteral(likePattern.get()), null);
-            predicate = logicalAnd(predicate, likePredicate);
-        }
-
-        Query query = simpleQuery(
-                selectList(aliasedName("table_name", "Table")),
-                from(catalogName, TABLE_TABLES),
-                predicate,
-                ordering(ascending("table_name")));
-
-        return process(query, context);
-    }
-
-    @Override
-    protected RelationType visitShowSchemas(ShowSchemas node, AnalysisContext context)
-    {
-        if (!node.getCatalog().isPresent() && !session.getCatalog().isPresent()) {
-            throw new SemanticException(CATALOG_NOT_SPECIFIED, node, "Catalog must be specified when session catalog is not set");
-        }
-
-        Optional<Expression> predicate = Optional.empty();
-        Optional<String> likePattern = node.getLikePattern();
-        if (likePattern.isPresent()) {
-            predicate = Optional.of(new LikePredicate(nameReference("schema_name"), new StringLiteral(likePattern.get()), null));
-        }
-
-        Query query = simpleQuery(
-                selectList(aliasedName("schema_name", "Schema")),
-                from(node.getCatalog().orElseGet(() -> session.getCatalog().get()), TABLE_SCHEMATA),
-                predicate,
-                ordering(ascending("schema_name")));
-
-        return process(query, context);
-    }
-
-    @Override
-    protected RelationType visitShowCatalogs(ShowCatalogs node, AnalysisContext context)
-    {
-        List<Expression> rows = metadata.getCatalogNames().keySet().stream()
-                .map(name -> row(new StringLiteral(name)))
-                .collect(toList());
-
-        Optional<Expression> predicate = Optional.empty();
-        Optional<String> likePattern = node.getLikePattern();
-        if (likePattern.isPresent()) {
-            predicate = Optional.of(new LikePredicate(nameReference("Catalog"), new StringLiteral(likePattern.get()), null));
-        }
-
-        Query query = simpleQuery(
-                selectList(new AllColumns()),
-                aliased(new Values(rows), "catalogs", ImmutableList.of("Catalog")),
-                predicate,
-                ordering(ascending("Catalog")));
-
-        return process(query, context);
-    }
-
-    @Override
-    protected RelationType visitShowColumns(ShowColumns showColumns, AnalysisContext context)
-    {
-        QualifiedObjectName tableName = createQualifiedObjectName(session, showColumns, showColumns.getTable());
-
-        if (!metadata.getView(session, tableName).isPresent() &&
-                !metadata.getTableHandle(session, tableName).isPresent()) {
-            throw new SemanticException(MISSING_TABLE, showColumns, "Table '%s' does not exist", tableName);
-        }
-
-        Query query = simpleQuery(
-                selectList(
-                        aliasedName("column_name", "Column"),
-                        aliasedName("data_type", "Type"),
-                        aliasedNullToEmpty("comment", "Comment")),
-                from(tableName.getCatalogName(), TABLE_COLUMNS),
-                logicalAnd(
-                        equal(nameReference("table_schema"), new StringLiteral(tableName.getSchemaName())),
-                        equal(nameReference("table_name"), new StringLiteral(tableName.getObjectName()))),
-                ordering(ascending("ordinal_position")));
-
-        return process(query, context);
-    }
-
-    private static <T> Expression getExpression(PropertyMetadata<T> property, Object value)
-            throws PrestoException
-    {
-        return toExpression(property.encode(property.getJavaType().cast(value)));
-    }
-
-    private static Expression toExpression(Object value)
-            throws PrestoException
-    {
-        if (value instanceof String) {
-            return new StringLiteral(value.toString());
-        }
-
-        if (value instanceof Boolean) {
-            return new BooleanLiteral(value.toString());
-        }
-
-        if (value instanceof Long || value instanceof Integer) {
-            return new LongLiteral(value.toString());
-        }
-
-        if (value instanceof Double) {
-            return new DoubleLiteral(value.toString());
-        }
-
-        if (value instanceof List) {
-            List<?> list = (List<?>) value;
-            return new ArrayConstructor(list.stream()
-                    .map(StatementAnalyzer::toExpression)
-                    .collect(toList()));
-        }
-
-        throw new PrestoException(INVALID_TABLE_PROPERTY, format("Failed to convert object of type %s to expression: %s", value.getClass().getName(), value));
-    }
-
-    @Override
     protected RelationType visitUse(Use node, AnalysisContext context)
     {
-        analysis.setUpdateType("USE");
         throw new SemanticException(NOT_SUPPORTED, node, "USE statement is not supported");
-    }
-
-    @Override
-    protected RelationType visitShowPartitions(ShowPartitions showPartitions, AnalysisContext context)
-    {
-        QualifiedObjectName table = createQualifiedObjectName(session, showPartitions, showPartitions.getTable());
-        Optional<TableHandle> tableHandle = metadata.getTableHandle(session, table);
-        if (!tableHandle.isPresent()) {
-            throw new SemanticException(MISSING_TABLE, showPartitions, "Table '%s' does not exist", table);
-        }
-
-        List<TableLayoutResult> layouts = metadata.getLayouts(session, tableHandle.get(), Constraint.alwaysTrue(), Optional.empty());
-        if (layouts.size() != 1) {
-            throw new SemanticException(NOT_SUPPORTED, showPartitions, "Table does not have exactly one layout: %s", table);
-        }
-        TableLayout layout = getOnlyElement(layouts).getLayout();
-        if (!layout.getDiscretePredicates().isPresent()) {
-            throw new SemanticException(NOT_SUPPORTED, showPartitions, "Table does not have partition columns: %s", table);
-        }
-        List<ColumnHandle> partitionColumns = layout.getDiscretePredicates().get().getColumns();
-
-            /*
-                Generate a dynamic pivot to output one column per partition key.
-                For example, a table with two partition keys (ds, cluster_name)
-                would generate the following query:
-
-                SELECT
-                  partition_number
-                , max(CASE WHEN partition_key = 'ds' THEN partition_value END) ds
-                , max(CASE WHEN partition_key = 'cluster_name' THEN partition_value END) cluster_name
-                FROM ...
-                GROUP BY partition_number
-
-                The values are also cast to the type of the partition column.
-                The query is then wrapped to allow custom filtering and ordering.
-            */
-
-        ImmutableList.Builder<SelectItem> selectList = ImmutableList.builder();
-        ImmutableList.Builder<SelectItem> wrappedList = ImmutableList.builder();
-        selectList.add(unaliasedName("partition_number"));
-        for (ColumnHandle columnHandle : partitionColumns) {
-            ColumnMetadata column = metadata.getColumnMetadata(session, tableHandle.get(), columnHandle);
-            Expression key = equal(nameReference("partition_key"), new StringLiteral(column.getName()));
-            Expression value = caseWhen(key, nameReference("partition_value"));
-            value = new Cast(value, column.getType().getTypeSignature().toString());
-            Expression function = functionCall("max", value);
-            selectList.add(new SingleColumn(function, column.getName()));
-            wrappedList.add(unaliasedName(column.getName()));
-        }
-
-        Query query = simpleQuery(
-                selectAll(selectList.build()),
-                from(table.getCatalogName(), TABLE_INTERNAL_PARTITIONS),
-                Optional.of(logicalAnd(
-                        equal(nameReference("table_schema"), new StringLiteral(table.getSchemaName())),
-                        equal(nameReference("table_name"), new StringLiteral(table.getObjectName())))),
-                Optional.of(new GroupBy(false, ImmutableList.of(new SimpleGroupBy(ImmutableList.of(nameReference("partition_number")))))),
-                Optional.empty(),
-                ImmutableList.of(),
-                Optional.empty());
-
-        query = simpleQuery(
-                selectAll(wrappedList.build()),
-                subquery(query),
-                showPartitions.getWhere(),
-                Optional.empty(),
-                Optional.empty(),
-                ImmutableList.<SortItem>builder()
-                        .addAll(showPartitions.getOrderBy())
-                        .add(ascending("partition_number"))
-                        .build(),
-                showPartitions.getLimit());
-
-        return process(query, context);
-    }
-
-    @Override
-    protected RelationType visitShowCreate(ShowCreate node, AnalysisContext context)
-    {
-        QualifiedObjectName objectName = createQualifiedObjectName(session, node, node.getName());
-        Optional<ViewDefinition> viewDefinition = metadata.getView(session, objectName);
-
-        if (node.getType() == VIEW) {
-            if (!viewDefinition.isPresent()) {
-                if (metadata.getTableHandle(session, objectName).isPresent()) {
-                    throw new SemanticException(NOT_SUPPORTED, node, "Relation '%s' is a table, not a view", objectName);
-                }
-                throw new SemanticException(MISSING_TABLE, node, "View '%s' does not exist", objectName);
-            }
-
-            Query query = parseView(viewDefinition.get().getOriginalSql(), objectName, node);
-            String sql = formatSql(new CreateView(createQualifiedName(objectName), query, false)).trim();
-            return process(singleValueQuery("Create View", sql), context);
-        }
-
-        if (node.getType() == TABLE) {
-            if (viewDefinition.isPresent()) {
-                throw new SemanticException(NOT_SUPPORTED, node, "Relation '%s' is a view, not a table", objectName);
-            }
-
-            Optional<TableHandle> tableHandle = metadata.getTableHandle(session, objectName);
-            if (!tableHandle.isPresent()) {
-                throw new SemanticException(MISSING_TABLE, node, "Table '%s' does not exist", objectName);
-            }
-
-            ConnectorTableMetadata connectorTableMetadata = metadata.getTableMetadata(session, tableHandle.get()).getMetadata();
-
-            List<TableElement> columns = connectorTableMetadata.getColumns().stream()
-                    .filter(column -> !column.isHidden())
-                    .map(column -> new TableElement(column.getName(), column.getType().getDisplayName()))
-                    .collect(toImmutableList());
-
-            Map<String, Object> properties = connectorTableMetadata.getProperties();
-            Map<String, PropertyMetadata<?>> allTableProperties = metadata.getTablePropertyManager().getAllTableProperties().get(objectName.getCatalogName());
-            Map<String, Expression> sqlProperties = new HashMap<>();
-
-            for (Map.Entry<String, Object> propertyEntry : properties.entrySet()) {
-                String propertyName = propertyEntry.getKey();
-                Object value = propertyEntry.getValue();
-                if (value == null) {
-                    throw new PrestoException(INVALID_TABLE_PROPERTY, format("Property %s for table %s cannot have a null value", propertyName, objectName));
-                }
-
-                PropertyMetadata<?> property = allTableProperties.get(propertyName);
-                if (!property.getJavaType().isInstance(value)) {
-                    throw new PrestoException(INVALID_TABLE_PROPERTY, format(
-                            "Property %s for table %s should have value of type %s, not %s",
-                            propertyName,
-                            objectName,
-                            property.getJavaType().getName(),
-                            value.getClass().getName()));
-                }
-
-                Expression sqlExpression = getExpression(property, value);
-                sqlProperties.put(propertyName, sqlExpression);
-            }
-
-            CreateTable createTable = new CreateTable(QualifiedName.of(objectName.getCatalogName(), objectName.getSchemaName(), objectName.getObjectName()), columns, false, sqlProperties);
-            Query query = singleValueQuery("Create Table", formatSql(createTable).trim());
-
-            return process(query, context);
-        }
-
-        throw new UnsupportedOperationException("SHOW CREATE only supported for tables and views");
-    }
-
-    @Override
-    protected RelationType visitShowFunctions(ShowFunctions node, AnalysisContext context)
-    {
-        ImmutableList.Builder<Expression> rows = ImmutableList.builder();
-        for (SqlFunction function : metadata.listFunctions()) {
-            if (function.getSignature().getKind() == APPROXIMATE_AGGREGATE) {
-                continue;
-            }
-            rows.add(row(
-                    new StringLiteral(function.getSignature().getName()),
-                    new StringLiteral(function.getSignature().getReturnType().toString()),
-                    new StringLiteral(Joiner.on(", ").join(function.getSignature().getArgumentTypes())),
-                    new StringLiteral(getFunctionType(function)),
-                    function.isDeterministic() ? TRUE_LITERAL : FALSE_LITERAL,
-                    new StringLiteral(nullToEmpty(function.getDescription()))));
-        }
-
-        Map<String, String> columns = ImmutableMap.<String, String>builder()
-                .put("function_name", "Function")
-                .put("return_type", "Return Type")
-                .put("argument_types", "Argument Types")
-                .put("function_type", "Function Type")
-                .put("deterministic", "Deterministic")
-                .put("description", "Description")
-                .build();
-
-        Query query = simpleQuery(
-                selectAll(columns.entrySet().stream()
-                        .map(entry -> aliasedName(entry.getKey(), entry.getValue()))
-                        .collect(toImmutableList())),
-                aliased(new Values(rows.build()), "functions", ImmutableList.copyOf(columns.keySet())),
-                ordering(
-                        ascending("function_name"),
-                        ascending("return_type"),
-                        ascending("argument_types"),
-                        ascending("function_type")));
-
-        return process(query, context);
-    }
-
-    private static String getFunctionType(SqlFunction function)
-    {
-        FunctionKind kind = function.getSignature().getKind();
-        switch (kind) {
-            case AGGREGATE:
-            case APPROXIMATE_AGGREGATE:
-                return "aggregate";
-            case WINDOW:
-                return "window";
-            case SCALAR:
-                return "scalar";
-        }
-        throw new IllegalArgumentException("Unsupported function kind: " + kind);
-    }
-
-    @Override
-    protected RelationType visitShowSession(ShowSession node, AnalysisContext context)
-    {
-        ImmutableList.Builder<Expression> rows = ImmutableList.builder();
-        List<SessionPropertyValue> sessionProperties = metadata.getSessionPropertyManager().getAllSessionProperties(session);
-        for (SessionPropertyValue sessionProperty : sessionProperties) {
-            if (sessionProperty.isHidden()) {
-                continue;
-            }
-
-            String value = sessionProperty.getValue();
-            String defaultValue = sessionProperty.getDefaultValue();
-            rows.add(row(
-                    new StringLiteral(sessionProperty.getFullyQualifiedName()),
-                    new StringLiteral(nullToEmpty(value)),
-                    new StringLiteral(nullToEmpty(defaultValue)),
-                    new StringLiteral(sessionProperty.getType()),
-                    new StringLiteral(sessionProperty.getDescription()),
-                    TRUE_LITERAL));
-        }
-
-        // add bogus row so we can support empty sessions
-        StringLiteral empty = new StringLiteral("");
-        rows.add(row(empty, empty, empty, empty, empty, FALSE_LITERAL));
-
-        Query query = simpleQuery(
-                selectList(
-                        aliasedName("name", "Name"),
-                        aliasedName("value", "Value"),
-                        aliasedName("default", "Default"),
-                        aliasedName("type", "Type"),
-                        aliasedName("description", "Description")),
-                aliased(
-                        new Values(rows.build()),
-                        "session",
-                        ImmutableList.of("name", "value", "default", "type", "description", "include")),
-                nameReference("include"));
-
-        return process(query, context);
     }
 
     @Override
@@ -859,16 +406,6 @@ class StatementAnalyzer
                 throw new SemanticException(NOT_SUPPORTED, node, "EXPLAIN ANALYZE only supports TYPE DISTRIBUTED option");
             }
             process(node.getStatement(), context);
-            Statement statement = analysis.getStatement();
-            // Some statements, like SHOW COLUMNS, are rewritten into a SELECT
-            if (statement != node.getStatement()) {
-                if (node.getLocation().isPresent()) {
-                    node = new Explain(node.getLocation().get(), node.isAnalyze(), statement, node.getOptions());
-                }
-                else {
-                    node = new Explain(statement, node.isAnalyze(), node.getOptions());
-                }
-            }
             analysis.setStatement(node);
             analysis.setUpdateType(null);
             RelationType type = new RelationType(Field.newUnqualified("Query Plan", VARCHAR));
@@ -1903,11 +1440,11 @@ class StatementAnalyzer
                     .collect(toImmutableList());
 
             for (Expression expression : Iterables.concat(outputExpressions, orderByExpressions)) {
-                verifyAggregations(node, distinctGroupingColumns, tupleDescriptor, expression, columnReferences);
+                verifyAggregations(distinctGroupingColumns, tupleDescriptor, expression, columnReferences);
             }
 
             if (node.getHaving().isPresent()) {
-                verifyAggregations(node, distinctGroupingColumns, tupleDescriptor, node.getHaving().get(), columnReferences);
+                verifyAggregations(distinctGroupingColumns, tupleDescriptor, node.getHaving().get(), columnReferences);
             }
         }
     }
@@ -1936,7 +1473,6 @@ class StatementAnalyzer
     }
 
     private void verifyAggregations(
-            QuerySpecification node,
             List<Expression> groupByExpressions,
             RelationType tupleDescriptor,
             Expression expression,
@@ -2106,10 +1642,5 @@ class StatementAnalyzer
         }
 
         analysis.setOrderByExpressions(node, orderByFieldsBuilder.build());
-    }
-
-    private static Relation from(String catalog, SchemaTableName table)
-    {
-        return table(QualifiedName.of(catalog, table.getSchemaName(), table.getTableName()));
     }
 }
